+++ src/command/list/mod.rs
@@ -51,6 +51,7 @@ pub use list_write::lrem;
 pub use list_write::lset;
 pub use list_write::ltrim;
 pub use list_write::rpop;
+pub use list_write::rpoplpush;
 pub use list_write::rpush;
 pub use list_write::rpushx;